# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Wegen NFT Trait Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /locker/{walletAddress}:
    get:
      operationId: getLocker
      tags: [locker]
      summary: Get a user's trait locker
      parameters:
        - name: walletAddress
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User's locker contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LockerResponse"
  /locker/{walletAddress}/purchase:
    post:
      operationId: purchaseTrait
      tags: [locker]
      summary: Record a trait purchase into user locker
      parameters:
        - name: walletAddress
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseTraitBody"
      responses:
        "201":
          description: Trait purchased and added to locker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LockerItem"
        "400":
          description: Invalid request or out of stock
...
  /bundles/{bundleId}/purchase:
    post:
      operationId: purchaseBundle
      tags: [bundles]
      summary: Purchase a trait bundle with ETH, adding all its traits to the wallet's locker
      parameters:
        - name: bundleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseBundleBody"
      responses:
        "201":
          description: Bundle purchased and traits added to locker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseBundleResponse"
        "404":
          description: Bundle not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Bundle sold out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/bundles:
    get:
      operationId: listAdminBundles
      tags: [admin]
      summary: List all trait bundles, including inactive ones (admin only)
      responses:
        "200":
          description: List of trait bundles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleListResponse"
    post:
      operationId: createBundle
      tags: [admin]
      summary: Create a new trait bundle (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBundleBody"
      responses:
        "201":
          description: Bundle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bundle"
  /admin/bundles/{bundleId}:
    put:
      operationId: updateBundle
      tags: [admin]
      summary: Update a trait bundle (admin only)
      parameters:
        - name: bundleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBundleBody"
      responses:
        "200":
          description: Bundle updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bundle"
    delete:
      operationId: deleteBundle
      tags: [admin]
      summary: Delete a trait bundle (admin only)
      parameters:
        - name: bundleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
components:
  schemas:
...
    PayoutSplit:
      type: object
      properties:
        walletAddress:
          type: string
          description: Ethereum wallet address receiving funds
        percentage:
          type: number
          minimum: 0
          maximum: 100
          description: Percentage of sale proceeds (0-100)
      required:
        - walletAddress
        - percentage
    Trait:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        theme:
          type: string
          nullable: true
          description: Named collection/theme this trait belongs to (e.g. "Stoner Traits", "70s Vibes")
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        mediaType:
          type: string
          nullable: true
        priceUsd:
          type: string
          description: Fixed USD price (source of truth, as string to avoid floating point issues)
        priceEth:
          type: string
          description: Live-computed ETH equivalent, snapshotted from priceUsd on last save (as string)
        priceWei:
          type: string
          description: Price in wei
        totalSupply:
          type: integer
        remainingSupply:
          type: integer
        isActive:
          type: boolean
        rarity:
          type: string
          enum: [common, uncommon, rare, legendary]
        nftCollection:
          type: string
          nullable: true
        payoutSplits:
          type: array
          items:
            $ref: "#/components/schemas/PayoutSplit"
          description: Wallet addresses and their percentage share of sale proceeds
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - priceUsd
        - priceEth
        - priceWei
        - totalSupply
        - remainingSupply
        - isActive
        - rarity
        - payoutSplits
        - createdAt
...
    LockerItem:
      type: object
      properties:
        id:
          type: integer
        traitId:
          type: integer
        walletAddress:
          type: string
        trait:
          $ref: "#/components/schemas/Trait"
        quantity:
          type: integer
        equippedToTokenId:
          type: integer
          nullable: true
        purchasedAt:
          type: string
          format: date-time
        txHash:
          type: string
          nullable: true
      required:
        - id
        - traitId
        - walletAddress
        - trait
        - quantity
        - equippedToTokenId
        - purchasedAt
    LockerResponse:
      type: object
      properties:
        walletAddress:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/LockerItem"
        totalItems:
          type: integer
      required:
        - walletAddress
        - items
        - totalItems
    PurchaseTraitBody:
      type: object
      properties:
        traitId:
          type: integer
        txHash:
          type: string
          nullable: true
        quantity:
          type: integer
          default: 1
      required:
        - traitId
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Bundle:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        priceUsd:
          type: string
          description: Fixed USD price (source of truth, as string to avoid floating point issues)
        priceEth:
          type: string
          description: Live-computed ETH equivalent, snapshotted from priceUsd on last save (as string)
        priceWei:
          type: string
        totalSupply:
          type: integer
        remainingSupply:
          type: integer
        isActive:
          type: boolean
        traits:
          type: array
          items:
            $ref: "#/components/schemas/Trait"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - priceUsd
        - priceEth
        - priceWei
        - totalSupply
        - remainingSupply
        - isActive
        - traits
        - createdAt
    BundleListResponse:
      type: object
      properties:
        bundles:
          type: array
          items:
            $ref: "#/components/schemas/Bundle"
      required:
        - bundles
    CreateBundleBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        priceUsd:
          type: string
          description: Fixed USD price; the ETH equivalent is derived server-side from the live rate on save
        totalSupply:
          type: integer
          default: -1
        isActive:
          type: boolean
          default: true
        traitIds:
          type: array
          items:
            type: integer
          description: IDs of traits included in this bundle
      required:
        - name
        - priceUsd
        - traitIds
    PurchaseBundleBody:
      type: object
      properties:
        walletAddress:
          type: string
        txHash:
          type: string
          nullable: true
      required:
        - walletAddress
    PurchaseBundleResponse:
      type: object
      properties:
        bundleId:
          type: integer
        walletAddress:
          type: string
        lockerItemIds:
          type: array
          items:
            type: integer
      required:
        - bundleId
        - walletAddress
        - lockerItemIds