# lib/api-zod/src/generated/types/bundle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wegen NFT Trait Store API
 * OpenAPI spec version: 0.1.0
 */
import type { Trait } from "./trait";

export interface Bundle {
  id: number;
  name: string;
  description?: string | null;
  imageUrl?: string | null;
  /** Fixed USD price (source of truth, as string to avoid floating point issues) */
  priceUsd: string;
  /** Live-computed ETH equivalent, snapshotted from priceUsd on last save (as string) */
  priceEth: string;
  priceWei: string;
  totalSupply: number;
  remainingSupply: number;
  isActive: boolean;
  traits: Trait[];
  createdAt: Date;
}