# lib/api-zod/src/generated/types/createBundleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wegen NFT Trait Store API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBundleBody {
  name: string;
  description?: string;
  imageUrl?: string;
  /** Fixed USD price; the ETH equivalent is derived server-side from the live rate on save */
  priceUsd: string;
  totalSupply?: number;
  isActive?: boolean;
  /** IDs of traits included in this bundle */
  traitIds: number[];
}