# lib/api-zod/src/generated/types/trait.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wegen NFT Trait Store API
 * OpenAPI spec version: 0.1.0
 */
import type { PayoutSplit } from "./payoutSplit";
import type { TraitRarity } from "./traitRarity";

export interface Trait {
  id: number;
  name: string;
  category: string;
  /** Named collection/theme this trait belongs to (e.g. "Stoner Traits", "70s Vibes") */
  theme?: string | null;
  description?: string | null;
  imageUrl?: string | null;
  mediaType?: string | null;
  /** Fixed USD price (source of truth, as string to avoid floating point issues) */
  priceUsd: string;
  /** Live-computed ETH equivalent, snapshotted from priceUsd on last save (as string) */
  priceEth: string;
  /** Price in wei */
  priceWei: string;
  totalSupply: number;
  remainingSupply: number;
  isActive: boolean;
  rarity: TraitRarity;
  nftCollection?: string | null;
  /** Wallet addresses and their percentage share of sale proceeds */
  payoutSplits: PayoutSplit[];
  createdAt: Date;
}